Prevent config resolution from throwing exceptions
